feat: migrate ownable contracts to Host ABI v1 with CBOR wire format
- Replace serde_cbor 0.11 with ciborium 0.2 across all contracts and ownable-std
  (serde_cbor fails to deserialize cbor-x 1.6.4's 2-byte map headers for enums)
- Rewrite worker.js to use ptr/len memory ABI (ownable_alloc/free/instantiate/execute/query/external_event)
- Add cbor-x encoding/decoding in WorkerRPC for all contract calls
- Fix externalEvent passing info correctly (was always undefined)
- Stop retrying apply chain on error to prevent error dialog loop
- Add TypedPackage ABI fields; update Builder/Package/Ownable services for new ABI

+++ ownables/basic/src/state.rs
@@ -11,4 +11,4 @@ pub const OWNABLE_INFO: Item<OwnableInfo> = Item::new("ownable_info");
 pub const METADATA: Item<Metadata> = Item::new("metadata");
 pub const NFT_ITEM: Item<NFT> = Item::new("nft");
 pub const PACKAGE_CID: Item<String> = Item::new("package_cid");
-pub const NETWORK_ID: Item<u8> = Item::new("network_id");
+pub const NETWORK_ID: Item<u32> = Item::new("network_id");
